+++ src/main/kotlin/racingcar/Car.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 class Car(private var position: Position = Position(0), private val moveable: Moveable = Moveable.Random()) {
     fun driving(): Position {
