+++ src/main/kotlin/racingcar/Moveable.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 import java.util.concurrent.ThreadLocalRandom
