+++ src/main/kotlin/racingcar/Positions.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 data class Positions(private val positions: List<Position>) : List<Position> by positions {
     constructor(vararg positions: Int) : this(positions.map(::Position))
