+++ src/main/kotlin/racingcar/UserInput.kt
@@ -1,4 +1,4 @@
-package step3
+package racingcar
 
 import java.io.InputStream
 import java.util.Scanner
